Separated legacy interface to its own file.
