+++ meteor-app/imports/api/connectors/postgresql/index.ts
@@ -1,1 +1,2 @@
 export * from "./postgresql";
+export * from "./legacy";
